# Add per GPU metrics in the output table