--- a/Dockerfile
+++ b/Dockerfile
@@ -30,6 +30,8 @@ ENV DCGM_VERSION=2.0.13
 # Ensure apt-get won't prompt for selecting options
 ENV DEBIAN_FRONTEND=noninteractive
 
+RUN apt update && apt install -y python3-dev
+
 # Install tritonclient from the SDK container
 COPY --from=sdk /workspace/install/python /tmp/tritonclient
 RUN find /tmp/tritonclient -maxdepth 1 -type f -name \