--- a/model_analyzer/record/record_aggregator.py
+++ b/model_analyzer/record/record_aggregator.py
@@ -11,15 +11,16 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from model_analyzer.record.record import Record
 from collections import defaultdict
+import itertools
 
 
 class RecordAggregator:
     """
     Stores a collection of Record objects.
     """
-
     def __init__(self):
         self._records = defaultdict(list)
 
@@ -40,10 +41,24 @@ class RecordAggregator:
             raise TritonModelAnalyzerException(
                 "Can only add objects of type 'Record' to RecordAggregator")
 
+    def add_key(self, record_type, records):
+        """
+        Adds or replaces all the records of a given record_type with the new
+        records
+
+        Parameters
+        ----------
+        record_type : Record
+            record_type to add to the records.
+        records : list
+            List of new records to be added.
+        """
+
+        self._records[record_type] = records
+
     def filter_records(self, record_types=None, filters=None):
         """
-        Get records that satisfy
-        the given list of criteria.
+        Get records that satisfy the given list of criteria.
 
         Parameters
         ----------
@@ -63,16 +78,23 @@ class RecordAggregator:
 
         Returns
         -------
-        Dict of List of Records
-            keys are record_types, values
-            are all records of that type
+        RecordAggregator
+            Returns a new RecordAggregator containing the filtered
+            records
         """
 
+        filtered_records = RecordAggregator()
         if not record_types and not filters:
-            return self._records
+            for record_type, records in self._records.items():
+                filtered_records.add_key(record_type, records)
+            return filtered_records
+
         if record_types and not filters:
             try:
-                return {k: self._records[k] for k in record_types}
+                for record_type in record_types:
+                    filtered_records.add_key(record_type,
+                                             self._records[record_type])
+                return filtered_records
             except KeyError as k:
                 raise TritonModelAnalyzerException(
                     f"Record type '{k.header()}' not found in this RecordAggregator"
@@ -87,13 +109,51 @@ class RecordAggregator:
             )
 
         # Remove records that do not satisfy criteria
-        filtered_records = defaultdict(list)
         for h, f in zip(record_types, filters):
-            filtered_records[h] = [
-                record for record in self._records[h] if f(record.value())
-            ]
+            for record in self._records[h]:
+                if f(record):
+                    filtered_records.insert(record)
 
         return filtered_records
+
+    def groupby(self, record_type, groupby_criteria, reduce_func=max):
+        """
+        Group all the records of a certain type together if they have the
+        same value for a given groupbby criteria.
+
+        Parameters
+        ----------
+        record_type : Record
+            A record type
+        groupby_criteria : callable
+            This callable will receive a single record as the argument and
+            must return the value that will be used for groupby
+        reduce_func : callable
+            A function to apply on the group of records
+
+        Returns
+        -------
+        dict
+            A dictionary where the keys are the unique values returned by
+            groupby_criteria and the values are the aggregated records.
+        """
+
+        records = self.filter_records(record_types=[record_type]).get_records()
+        field_values = set()
+        for record in records[record_type]:
+            field_values.add(groupby_criteria(record))
+
+        groupby_result = {}
+        for field_value in field_values:
+            aggregated_result = self.filter_records(
+                record_types=[record_type],
+                filters=[
+                    lambda record: groupby_criteria(record) == field_value
+                ]).aggregate(record_types=[record_type],
+                             reduce_func=reduce_func)
+            groupby_result[field_value] = aggregated_result[record_type]
+
+        return groupby_result
 
     def record_types(self):
         """
@@ -104,7 +164,7 @@ class RecordAggregator:
             RecordAgrregator
         """
 
-        return list(self._records.keys())
+        return list(self._records)
 
     def total(self, record_type=None):
         """
@@ -160,3 +220,23 @@ class RecordAggregator:
                 values.append(record.value())
             aggregated_records[record_type] = reduce_func(values)
         return aggregated_records
+
+    def get_records(self):
+        """
+        Get all the records.
+
+        Returns
+        -------
+        dict
+            A dictionary where the keys are record types and the values are
+            an array of records with the specified type
+        """
+
+        return self._records
+
+    def _flatten_records(self, records):
+        """
+        Flatten the records array by joining all the arrays together.
+        """
+
+        return list(itertools.chain.from_iterable(records))