Before
```
Server Only:
Model                         Batch   Concurrency   Throughput(infer/sec)   Max GPU Utilization(%)   Max GPU Used Memory(MB)   Max GPU Free Memory(MB)  
triton-server                 0       0             0                       0.0                      228.0                     23991.0                  

Models:
Model                         Batch   Concurrency   Throughput(infer/sec)   Max GPU Utilization(%)   Max GPU Used Memory(MB)   Max GPU Free Memory(MB)  
classification_chestxray_v1   1       1             49.2                    37.0                     5186.0                    23359.0                  
```

After
```
Server Only:
Model           Batch   Concurrency   Max GPU Utilization(%)   Max GPU Used Memory(MB)   Max GPU Free Memory(MB)   GPU ID  
triton-server   0       0             0.0                      319.0                     15961.0                   0       
triton-server   0       0             0.0                      319.0                     15961.0                   1       
triton-server   0       0             0.0                      319.0                     15961.0                   2       
triton-server   0       0             0.0                      319.0                     15961.0                   3       

Models (GPU Metrics):
Model                         Batch   Concurrency   Max GPU Utilization(%)   Max GPU Used Memory(MB)   Max GPU Free Memory(MB)   GPU ID  
classification_chestxray_v1   1       1             66.0                     4765.0                    15439.0                   0       
classification_chestxray_v1   1       1             54.0                     4729.0                    15439.0                   1       
classification_chestxray_v1   1       1             65.0                     4729.0                    15439.0                   2       
classification_chestxray_v1   1       1             64.0                     4729.0                    15439.0                   3       

Models (Inference):
Model                         Batch   Concurrency   Throughput(infer/sec)  
classification_chestxray_v1   1       1             56.0 
```