@dzier 2021-01-05
I think the GPU ID should be the second column instead of the last, since it is really part of the run identifier and not a resulting metric.