@Tabrizian 2021-01-06
> I think the GPU ID should be the second column instead of the last, since it is really part of the run identifier and not a resulting metric.
```
Models (GPU Metrics):
Model             GPU ID   Batch   Concurrency   Max GPU Utilization(%)   Max GPU Used Memory(MB)   Max GPU Free Memory(MB)  
vgg_16_graphdef   0        2       1             93.0                     7628.0                    23383.0                  

Server Only:
Model           GPU ID   Batch   Concurrency   Max GPU Utilization(%)   Max GPU Used Memory(MB)   Max GPU Free Memory(MB)  
triton-server   0        0       0             0.0                      228.0                     23991.0                  

Models (Inference):
Model             Batch   Concurrency   Throughput(infer/sec)  
vgg_16_graphdef   2       1             55.2 

```

@dzier